+++ lib/govuk_app_config/railtie.rb
@@ -1,0 +1,7 @@
+module GovukAppConfig
+  class Railtie < Rails::Railtie
+    config.before_initialize do
+      GovukLogging.configure if Rails.env.production?
+    end
+  end
+end