+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+* Add Unicorn (our web server) as a dependency. Make sure to drop `gem 'unicorn'` from your Gemfile after upgrading.
+
 ## 0.3.0
 
 * Add `time` and `gauge` to `GovukStatsd`